gestion des erreurs

+++ frontend/src/components/Users/DeleteAccount.jsx
@@ -1,10 +1,9 @@
-import React, {useState, useCallback} from 'react';
+import React, {useCallback} from 'react';
 import { Link } from 'react-router-dom';
 import AuthApi from '../AuthApi';
 import Cookies from 'js-cookie';
 
 function DeleteAccount () {
-    const [error, setError] = useState(null);
     const Auth = React.useContext(AuthApi);
 
     const storage = JSON.parse(localStorage.getItem('userConnect'));
@@ -26,31 +25,33 @@ function DeleteAccount () {
         })
         .then(res => res.json())
         .then(
-            () => {
-                Auth.setAuth(false);
-                Cookies.remove("user");
-                localStorage.clear();
-            },
-            (error) => {
-                setError(error);
+            (res) => {
+                if (res.ok) { 
+                    alert("Compte supprimé !")
+                    Auth.setAuth(false);
+                    Cookies.remove("user");
+                    localStorage.clear();
+                } else { 
+                    alert("Votre compte n'a pas pu être supprimé."); 
+                }
             }
         )
+        .catch(error => {
+            this.setState({ Erreur: error.toString() });
+            alert("Votre compte n'a pas pu être supprimé !");
+            console.error('There was an error!', error);
+        })
     }, [Auth, userId, token])
 
-    if (error) {
-        return <div>Erreur : {error.message}</div>;
-    } else {
-        return (<>
-            <div className="container">
-                <h1>Souhaitez vous vraiment supprimer votre compte ?</h1>
-                <div className="form-submit">
-                    <Link to={'/user/' + userId} className="btn btn-outline-info btn-sm">Retour à mon compte</Link>
-                    <button className="btn btn-outline-danger btn-sm" onClick={handleSubmit}>Supprimer mon compte</button>
-                </div>
+    return (
+        <div className="container">
+            <h1>Souhaitez vous vraiment supprimer votre compte ?</h1>
+            <div className="form-submit">
+                <Link to={'/user/' + userId} className="btn btn-outline-info btn-sm">Retour à mon compte</Link>
+                <button className="btn btn-outline-danger btn-sm" onClick={handleSubmit}>Supprimer mon compte</button>
             </div>
-        </>
-        );
-    } 
+        </div>
+    );
 }
 
 export default DeleteAccount;
